+++ database/factories/ModelFactory.php
@@ -49,6 +49,9 @@ $factory->define(\App\Item::class, function (Faker\Generator $faker) {
         'inscription' => $faker->word,
         'related_work_order' => $faker->randomNumber,
         'related_work_total' => $faker->randomNumber,
+        'colors' => [
+            $faker->hexColor => 1,
+        ]
     ];
 });