+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file[^1].
 - upgrade laravel translatable models package
 - rewrite and remove bouncy package usages
 - refactor authority/item filters to use symfony forms
+- range color search
 
 ## [1.9.1] - 2019-07-30
 ### Changed